// src/database/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Type = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE'
} as const

export type Type = (typeof Type)[keyof typeof Type]


export const TicketStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]